+++ App.js
@@ -1,30 +1,22 @@
 import React from 'react';
 import styled from 'styled-components/native';
-import { SensorComponent } from './components/SensorComponent';
 import { createDrawerNavigator } from '@react-navigation/drawer';
 import { NavigationContainer } from '@react-navigation/native';
 
+import Dice from './components/Dice';
+import EntryPage from './components/EntryPage';
+
 const Drawer = createDrawerNavigator();	
 
-const Container = styled.View`
-	flex: 1;
-	background-color: papayawhip;
-	justify-content: center;
-	align-items: center;
-`;
-
-const Title = styled.Text`
-	font-size: 24px;
-	color: palevioletred;
-`;
 
 const App = () => {
-	return (
-		<NavigationContainer>
-		<Drawer.Navigator>
-		<Drawer.Screen name="Start" component={SensorComponent} />
-		</Drawer.Navigator>
-		</NavigationContainer>
+	return (		
+	<NavigationContainer>
+		<Drawer.Navigator initialRouteName="Home">
+			<Drawer.Screen name="Start" component={EntryPage} />
+			<Drawer.Screen name="Dice" component={Dice} />
+		</Drawer.Navigator>	
+	</NavigationContainer>
 	);
 };
